Clean up of some code in `restrictions.py`
